Patch by mintaka (http://www.fractalforums.com/mandelbulber/mandelbulber-1-06-patch/)
The fix i made addresses an issue with navigation. It only reveals when you let it render an animation. If you turn the camera between keyframes with 3-axis rotation and one of the angles goes above 180° or below -180°, then the animation makes weird turns. This is not a bug but it's annoying to have to correct the angles manually. The reason for this is the math that is used to calculate the new set of view angles, which involves arcsin and arctan functions. The nature of these functions is that they are repetitive below -180° and above +180°. So, if you make for example a turn from 170° to 190°, it gives -170° as a result, which is equivalent to +190° in math but not the same for the frame interpolation - then the camera makes almost a full revolution between two keyframes (340°).
This patch fixes the view, it should work with turns up to 180° but it doesn't do much math, so 180 is a safe guess where it should always recognize the error.

The second change in this patch is the addition of two navigation buttons: rotate the camera around its axis (gamma rotation). I made icons for the buttons that fit the style of the existing ones.